ingress/pyproject.toml
[project]
name = "py_poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.13",
    "nats-py>=2.10.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
    "sqlalchemy[asyncio]>=2.0.41",
    "sqlmodel>=0.0.24",
]
